third_party/WebKit/LayoutTests/platform/win/webaudio/AudioParam/audioparam-clamp-time-to-current-time-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = 2 duplicate test names: "Output[0-127] contains only the constant 1.", "Output[128-1023] contains only the constant 0.5."
PASS Output[0-127] contains only the constant 0. 
PASS Output[128-1023] contains only the constant 1. 
PASS *** setValueAtTime in the past correctly clamped to current time. 
PASS Output[0-127] contains only the constant 1. 
PASS Output[128-1023] contains only the constant 0.5. 
PASS *** linearRampToValueAtTime in the past correctly clamped to current time. 
PASS Output[0-127] contains only the constant 1. 
PASS Output[128-1023] contains only the constant 0.5. 
PASS *** exponentialRampToValueAtTime in the past correctly clamped to current time. 
PASS Output[0-127] contains only the constant 1. 
PASS Output[128] is equal to 1. 
PASS Output[129-1023] contains only positive values is equal to true. 
PASS *** setTargetAtTime in the past correctly clamped to current time. 
PASS Output[0-127] contains only the constant 1. 
PASS Output[128-1023] is equal to true. 
PASS *** setValueCurveAtTime in the past correctly clamped to current time. 
Harness: the test ran to completion.
